requirements.txt
fastapi~=0.115.11
pydantic~=2.10.6
requests~=2.32.3
spotipy~=2.25.1
python-dotenv~=1.0.1